+++ cmd/smpe_lint/main.go
@@ -15,7 +15,7 @@ import (
 )
 
 var (
-	version = "v1.3.4"
+	version = "v1.3.5"
 	commit  = "unknown"
 )