+++ cmd/smpe_outl/main.go
@@ -14,7 +14,7 @@ import (
 )
 
 var (
-	version = "v1.3.4"
+	version = "v1.3.5"
 	commit  = "unknown"
 )